Update README.md
The Online Examination System is a Java-based web application designed to facilitate the conduct of online exams. It provides an intuitive and secure platform for both administrators and examinees to create, manage, and take exams respectively. This project aims to streamline the examination process, eliminate manual grading tasks, and provide a convenient way for users to assess and evaluate knowledge and skills.